Add lenient configuration
This allows you to specify a lenient mode that capturing a returned disposable is sort of a "I know what I'm doing" mode
